+++ crates/symbolicator/src/endpoints/mod.rs
@@ -1,28 +1,46 @@
-use actix_web::http::StatusCode;
-use actix_web::{App, HttpResponse, ResponseError};
+use axum::handler::{get, post};
+use axum::Router;
+use sentry::integrations::tower::NewSentryLayer;
 
-use crate::services::symbolication::MaxRequestsError;
+use crate::metrics::MetricsLayer;
 use crate::services::Service;
+use crate::utils::sentry::SentryRequestLayer;
 
 mod applecrashreport;
-mod healthcheck;
+mod error;
 mod minidump;
+mod multipart;
 mod proxy;
 mod requests;
 mod symbolicate;
 
-/// Adds all endpoint routes to the app.
-pub fn configure(app: App<Service>) -> App<Service> {
-    app.configure(applecrashreport::configure)
-        .configure(healthcheck::configure)
-        .configure(minidump::configure)
-        .configure(proxy::configure)
-        .configure(requests::configure)
-        .configure(symbolicate::configure)
+pub use error::ResponseError;
+
+use applecrashreport::handle_apple_crash_report_request as applecrashreport;
+use minidump::handle_minidump_request as minidump;
+use proxy::proxy_symstore_request as proxy;
+use requests::poll_request as requests;
+use symbolicate::symbolicate_frames as symbolicate;
+
+pub async fn healthcheck() -> &'static str {
+    metric!(counter("healthcheck") += 1);
+    "ok"
 }
 
-impl ResponseError for MaxRequestsError {
-    fn error_response(&self) -> actix_web::HttpResponse {
-        HttpResponse::new(StatusCode::SERVICE_UNAVAILABLE)
-    }
+pub type App = Router<axum::routing::BoxRoute>;
+
+pub fn create_app(service: Service) -> App {
+    Router::new()
+        .route("/proxy/:path", get(proxy).head(proxy))
+        .route("/requests/:request_id", get(requests))
+        .route("/symbolicate", post(symbolicate))
+        .route("/minidump", post(minidump))
+        .route("/applecrashreport", post(applecrashreport))
+        .layer(axum::AddExtensionLayer::new(service))
+        .layer(SentryRequestLayer)
+        .layer(NewSentryLayer::new_from_top())
+        .layer(MetricsLayer)
+        // the healthcheck is last, as it will bypass all the middlewares
+        .route("/healthcheck", get(healthcheck))
+        .boxed()
 }