+++ crates/symbolicator/src/server.rs
@@ -1,26 +1,15 @@
-use actix_web::{server::HttpServer, App};
+use std::net::SocketAddr;
+
 use anyhow::{Context, Result};
 
 use crate::config::Config;
 use crate::endpoints;
-use crate::middlewares;
 use crate::services::Service;
-use crate::utils::sentry::SentryMiddleware;
-
-/// Creates the Actix web application with all middlewares.
-#[inline]
-pub fn create_app(state: Service) -> App<Service> {
-    App::with_state(state)
-        .middleware(middlewares::Metrics)
-        .middleware(middlewares::ErrorHandlers)
-        .middleware(SentryMiddleware::new())
-        .configure(endpoints::configure)
-}
 
 /// Starts all actors and HTTP server based on loaded config.
 pub fn run(config: Config) -> Result<()> {
     // Log this metric before actually starting the server. This allows to see restarts even if
-    // service creation fails. The HTTP server is bound before the actix system runs.
+    // service creation fails.
     metric!(counter("server.starting") += 1);
 
     let megs = 1024 * 1024;
@@ -37,7 +26,7 @@ pub fn run(config: Config) -> Result<()> {
         .build()
         .unwrap();
 
-    let bind = config.bind.clone();
+    let socket = config.bind.parse::<SocketAddr>()?;
 
     let service = Service::create(
         config,
@@ -46,11 +35,14 @@ pub fn run(config: Config) -> Result<()> {
     )
     .context("failed to create service state")?;
 
-    log::info!("Starting http server: {}", bind);
-    HttpServer::new(move || create_app(service.clone()))
-        .bind(&bind)
-        .context("failed to bind to the port")?
-        .run();
+    let _guard = io_pool.enter();
+    let server =
+        axum::Server::bind(&socket).serve(endpoints::create_app(service).into_make_service());
+
+    log::info!("Starting server on {}", server.local_addr());
+
+    io_pool.block_on(server)?;
+
     log::info!("System shutdown complete");
 
     Ok(())